fix!: callback parameter in pushNotification
